+++ misc/update_DESCRIPTION_NEWS.R
@@ -378,7 +378,8 @@ my_news$add_bullet(c("changing error messages"))
 my_desc$bump_version("patch")
 my_news$add_version(my_desc$get_version())
 my_news$add_subtitle("Bugfixes")
-my_news$add_bullet(c("fixing plugin for coment line setting"))
+my_news$add_bullet(c("fixed plugin for coment line setting",
+                     "fixed FUN argument error in read_files"))
 
 # save everything ---------------------------------------------------------